chore: refactor operations files
