+++ client/app/(authenticated)/idir/cas_analyst/dashboard/operations/page.tsx
@@ -1,6 +1,6 @@
 // 🚩 flagging that for shared routes between roles, "Page" code is a component for code maintainability
+import { OperationsSearchParams } from "@/app/components/operations/types";
 import OperationsPage from "@/app/components/routes/operations/Page";
-import { OperationsSearchParams } from "@/app/components/routes/operations/types";
 
 export default async function Page({
   searchParams,
