+++ client/app/components/operations/types.ts
@@ -1,4 +1,4 @@
-import { Operator } from "../select-operator/form/types";
+import { Operator } from "../routes/select-operator/form/types";
 
 export interface Operation {
   id: number;
