--- client/app/components/routes/operations/create/Page.tsx
@@ -1,11 +1,0 @@
-import { Suspense } from "react";
-import Operation from "@/app/components/routes/operations/form/Operation";
-import Loading from "@/app/components/loading/SkeletonForm";
-
-export default async function OperationsCreatePage() {
-  return (
-    <Suspense fallback={<Loading />}>
-      <Operation />
-    </Suspense>
-  );
-}
