+++ client/app/components/routes/operations/operation/Page.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from "react";
-import Operation from "@/app/components/routes/operations/form/Operation";
 import Loading from "@/app/components/loading/SkeletonForm";
+import Operation from "@/app/components/operations/Operation";
 
 export default async function OperationsOperationPage({
   params,
